+++ src/vi/io/csv_file.cpp
@@ -20,7 +20,7 @@ void csv_file::load(vi::la::matrix& matrix, std::vector<std::string>& header) {
 
 void csv_file::load(vi::la::matrix& matrix, std::vector<std::string>* header) {
   _stream.exceptions(std::iostream::badbit);
-  std::vector<std::vector<double>> matrix_values;
+  std::vector<std::vector<float>> matrix_values;
   size_t max_columns(0U);
 
   bool first_line(true);
@@ -32,7 +32,7 @@ void csv_file::load(vi::la::matrix& matrix, std::vector<std::string>* header) {
       continue;
     }
 
-    std::vector<double> row;
+    std::vector<float> row;
     parse_row(line, row);
     max_columns = std::max(max_columns, row.size());
     matrix_values.push_back(row);
@@ -50,17 +50,17 @@ void csv_file::parse_header(const std::string& line, std::vector<std::string>& h
   }
 }
 
-void csv_file::parse_row(const std::string& line, std::vector<double>& row) const {
+void csv_file::parse_row(const std::string& line, std::vector<float>& row) const {
   std::istringstream row_stream(line);
   std::string value;
   while (std::getline(row_stream, value, _delimiter)) {
-    row.push_back(std::stod(value));
+    row.push_back(std::stof(value));
   }
 }
 
-std::shared_ptr<double> csv_file::make_buffer(const std::vector<std::vector<double>>& matrix_values,
+std::shared_ptr<float> csv_file::make_buffer(const std::vector<std::vector<float>>& matrix_values,
                                               size_t max_columns) const {
-  double* values = new double[matrix_values.size() * max_columns];
+  float* values = new float[matrix_values.size() * max_columns];
   size_t offset = 0U;
   for (auto row : matrix_values) {
     for (size_t column = 0U; column < row.size(); ++column) {
@@ -70,7 +70,7 @@ std::shared_ptr<double> csv_file::make_buffer(const std::vector<std::vector<doub
     offset += max_columns;
   }
 
-  return std::shared_ptr<double>(values, [](double* p) { delete[] p; });
+  return std::shared_ptr<float>(values, [](float* p) { delete[] p; });
 }
 
 void csv_file::store(const vi::la::matrix& matrix) { store(matrix, nullptr); }
